separate logout functions from auth.ts
