+++ src/components/IdleStatusMonitor.js
@@ -3,7 +3,7 @@ import _ from 'lodash/fp';
 import * as qs from 'qs';
 import { useEffect, useState } from 'react';
 import { div, h, iframe } from 'react-hyperscript-helpers';
-import { signOut } from 'src/auth/auth';
+import { signOut } from 'src/auth/auth-events/logout';
 import ButtonBar from 'src/components/ButtonBar';
 import colors from 'src/libs/colors';
 import * as Nav from 'src/libs/nav';
