+++ src/components/TopBar.js
@@ -6,7 +6,8 @@ import { a, div, h, h1, img, span } from 'react-hyperscript-helpers';
 import { Transition } from 'react-transition-group';
 import { AlertsIndicator } from 'src/alerts/Alerts';
 import { RequiredUpdateAlert } from 'src/alerts/RequiredUpdateAlert';
-import { signIn, signOut } from 'src/auth/auth';
+import { signIn } from 'src/auth/auth';
+import { signOut } from 'src/auth/auth-events/logout';
 import { Clickable, IdContainer, LabeledCheckbox, Link, spinnerOverlay } from 'src/components/common';
 import { icon } from 'src/components/icons';
 import { TextArea } from 'src/components/input';
