+++ src/registration/Register.test.ts
@@ -2,7 +2,8 @@ import { DeepPartial } from '@terra-ui-packages/core-utils';
 import { act, fireEvent, screen } from '@testing-library/react';
 import { axe } from 'jest-axe';
 import { h } from 'react-hyperscript-helpers';
-import { loadTerraUser, signOut } from 'src/auth/auth';
+import { loadTerraUser } from 'src/auth/auth';
+import { signOut } from 'src/auth/auth-events/logout';
 import { Ajax } from 'src/libs/ajax';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
 
@@ -10,10 +11,15 @@ import { Register } from './Register';
 
 jest.mock('src/libs/ajax');
 
+jest.mock('src/auth/auth-events/logout', () => ({
+  ...jest.requireActual('src/auth/auth-events/logout'),
+
+  signOut: jest.fn(),
+}));
+
 jest.mock('src/auth/auth', () => ({
   ...jest.requireActual('src/auth/auth'),
   loadTerraUser: jest.fn(),
-  signOut: jest.fn(),
 }));
 
 jest.mock('react-notifications-component', () => {
